requirements.txt
certifi==2016.9.26
django-redis-cache==1.7.1
requests==2.2.1
Django==1.10.4
future==0.16.0
gunicorn==19.6.0
dj-database-url==0.4.2
python-telegram-bot==5.3.0
urllib3==1.21.1
psycopg2==2.6.1
whitenoise==3.3.0